backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
docstring_parser==0.16
environs==11.0.0
fastapi==0.115.3
fastapi_cors==0.0.6
google-api-core==2.21.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.20.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
marshmallow==3.23.0
moviepy==1.0.3
numpy==2.1.2
packaging==24.1
pillow==11.0.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
rsa==4.9
schwabdev==2.3.0
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
starlette==0.41.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
vertexai==1.70.0
websockets==13.1
